Refactor : ReaderImplTest, StoreImplTest을 domain -> infra로 경로 이동
